remove more unused styles
