+++ app/components/Layout/utils/justifyContent.ts
@@ -1,0 +1,11 @@
+const justifyContent = ({ spaceAround = false, centered = false }) => {
+  if (spaceAround) {
+    return 'space-around';
+  }
+  if (centered) {
+    return 'center';
+  }
+  return 'flex-start';
+};
+
+export default justifyContent;
