--- app/components/Menu/index.ts
@@ -1,4 +1,0 @@
-import MenuItem from './MenuItem';
-
-// eslint-disable-next-line import/prefer-default-export
-export { MenuItem };
