--- app/modules/editor/components/EnduranceForm/EnduranceForm.styled.ts
@@ -1,22 +1,0 @@
-import styled from 'styled-components/native';
-
-export const Wrapper = styled.View`
-  flex: 1;
-  padding: 10px;
-`;
-
-export const MainWrapper = styled.View`
-  flex: 1;
-  justify-content: flex-start;
-`;
-
-export const Block = styled.View`
-  height: 150px;
-  justify-content: center;
-`;
-
-export const RowContainer = styled.View`
-  flex: 1;
-  flex-direction: row;
-  width: 100%;
-`;
