+++ app/modules/editor/components/StaticFormInputs/TableBaseInput.tsx
@@ -1,21 +1,15 @@
 import React, { FC } from 'react';
-import styled from 'styled-components/native';
+import { Stack } from '../../../../components/Layout';
 import LongTouchButton from '../../../../components/LongTouchButton';
 import { useAppDispatch } from '../../../../redux/hooks';
 import { changeTableBase } from '../../redux/actions/composed/changeTableBase';
 import { useEditorBaseSelector } from '../../redux/editorSelectors';
 
-export const ButtonsSet = styled.View`
-  flex: 1;
-  flex-direction: row;
-  align-items: center;
-`;
-
 const TableBaseInput: FC = () => {
   const dispatch = useAppDispatch();
   const base = useEditorBaseSelector();
   return (
-    <ButtonsSet>
+    <Stack horizontal>
       <LongTouchButton
         title="-"
         onPressStart={() => dispatch(changeTableBase(base - 5))}
@@ -26,7 +20,7 @@ const TableBaseInput: FC = () => {
         onPressStart={() => dispatch(changeTableBase(base + 5))}
         onPressInterval={() => dispatch(changeTableBase(base + 5))}
       />
-    </ButtonsSet>
+    </Stack>
   );
 };
