+++ app/modules/selector/SelectorScene.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import { AppScreenType, RoutesEnum } from '../../Routes';
 import { COLOR_DARK, COLOR_DARKER, COLOR_LIGHT, COLOR_NORMAL } from '../../commonStyles';
 import { ScrollableStack } from '../../components/Layout';
-import { MenuItem } from '../../components/Menu';
+import MenuItem from '../../components/MenuItem/MenuItem';
 import SceneWrapper from '../../components/SceneWrapper/SceneWrapper';
 import useAppNavitation from '../../useAppNavigation';
